able to batch deploy dex and token
