+++ batch-transactions/src/utils.ts
@@ -1,42 +1,38 @@
-// TODO: remove me and export them from bodhi
-import { BigNumber, BigNumberish } from '@ethersproject/bignumber';
-import { BytesLike } from '@ethersproject/bytes';
-import { bufferToU8a, hexToBn, isBuffer, isHex, isU8a, u8aToBn, u8aToHex } from '@polkadot/util';
-import BN from 'bn.js';
+import { ContractFactory } from 'ethers';
+import { Provider, Signer } from '@acala-network/bodhi';
+import { toBN } from './utils-from-bodhi';
 
-export const U32MAX = BigNumber.from('0xffffffff');
-export const U64MAX = BigNumber.from('0xffffffffffffffff');
+import dexContract from './Dex.json';
+import tokenContract from './Token.json';
 
-export function toBN(bigNumberis: BigNumberish = 0): BN {
-  if (isU8a(bigNumberis)) {
-    return u8aToBn(bigNumberis);
-  }
-  if (isHex(bigNumberis)) {
-    return hexToBn(bigNumberis);
-  }
+// TODO: add and export this util in bodhi
+export const getDeployExtrinsic = async (signer: Signer, abi: any, bytecode: string, args: any[] = []) => {
+  const factory = new ContractFactory(abi, bytecode, signer);
 
-  if (BigNumber.isBigNumber(bigNumberis)) {
-    const hex = bigNumberis.toHexString();
-    if (hex[0] === '-') {
-      return new BN('-' + hex.substring(3), 16);
-    }
-    return new BN(hex.substring(2), 16);
-  }
+  const evmAddress = await signer.getAddress();
+  const tx = await signer.populateTransaction({
+    ...factory.getDeployTransaction(...args),
+    from: evmAddress,
+  });
 
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  return new BN(bigNumberis as any);
-}
+  const {
+    gas: gasLimit,
+    storage: storageLimit,
+  } = await signer.provider.estimateResources(tx);
 
-export function dataToString(bytes: BytesLike): string {
-  if (isBuffer(bytes)) {
-    return u8aToHex(bufferToU8a(bytes));
-  }
-  if (isU8a(bytes)) {
-    return u8aToHex(bytes);
-  }
-  if (Array.isArray(bytes)) {
-    return u8aToHex(Buffer.from(bytes));
-  }
+  return signer.provider.api.tx.evm.create(
+    tx.data,
+    toBN(tx.value),
+    toBN(gasLimit),
+    toBN(storageLimit.isNegative() ? 0 : storageLimit),
+    tx.accessList || []
+  );
+};
 
-  return bytes as string;
-}
+export const getDexDeployExtrinsic = async (signer: Signer) => (
+  getDeployExtrinsic(signer, dexContract.abi, dexContract.bytecode)
+);
+
+export const getTokenDeployExtrinsic = async (signer: Signer, args: any[]) => (
+  getDeployExtrinsic(signer, tokenContract.abi, tokenContract.bytecode, args)
+);
